refactor: TemplateMaker를 web-server 모듈로 이동
