+++ web-server/src/main/java/utils/TemplateMaker.java
@@ -1,4 +1,4 @@
-package util;
+package utils;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Helper;
